backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
cryptography>=42.0.0
httpx>=0.27.0
mcp>=1.0.0
alembic>=1.13.0
python-jose>=3.3.0
openai>=1.0.0
anthropic>=0.39.0
# python-amazon-paapi>=6.0.0  # Optional: for product images by ASIN. Uncomment and pip install to enable.

# ── Test dependencies ────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio[trio]>=4.0.0
